requirements.txt
# Requirements file

pyserial
dependency_injector